base_refactor/refactor encoder, decoder and seq2seq modules
