--- configs/seq2seq_cnn.yaml
@@ -1,33 +1,0 @@
-model_name: seq2seq_cnn
-
-with_pad: True
-inversion_dataset: False
-
-general:
-  logs_dir: logs
-  ngpu: 0
-  cuda: True
-
-base_model:
-  enc_emb_dim: 256
-  enc_hid_dim: 512
-  enc_layers: 10
-  enc_kernel_size: 3
-  enc_dropout: 0.25
-  dec_emb_dim: 256
-  dec_hid_dim: 512
-  dec_layers: 10
-  dec_kernel_size: 3
-  dec_dropout: 0.25
-
-checkpoint_callback:
-  mode: min
-  monitor: val_loss
-  save_top_k: 3
-  verbose: True
-
-train_params:
-  epoch_size: 10
-  batch_size: 128
-  n_epochs: 10
-  clip: 1
