+++ machine_translation/attention/__init__.py
@@ -1,0 +1,1 @@
+from .gru_attention import Attention
