+++ machine_translation/attention/gru_attention.py
@@ -1,0 +1,44 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class Attention(nn.Module):
+    def __init__(
+            self,
+            enc_hid_dim,
+            dec_hid_dim
+    ):
+        super().__init__()
+
+        self.attn = nn.Linear((enc_hid_dim * 2) + dec_hid_dim, dec_hid_dim)
+        self.v = nn.Linear(dec_hid_dim, 1, bias=False)
+
+    def forward(
+            self,
+            hidden,
+            encoder_outputs
+    ):
+        # hidden = [batch size, dec hid dim]
+        # encoder_outputs = [src len, batch size, enc hid dim * 2]
+
+        batch_size = encoder_outputs.shape[1]
+        src_len = encoder_outputs.shape[0]
+
+        # repeat decoder hidden state src_len times
+        hidden = hidden.unsqueeze(1).repeat(1, src_len, 1)
+
+        encoder_outputs = encoder_outputs.permute(1, 0, 2)
+
+        # hidden = [batch size, src len, dec hid dim]
+        # encoder_outputs = [batch size, src len, enc hid dim * 2]
+
+        energy = torch.tanh(self.attn(torch.cat((hidden, encoder_outputs), dim=2)))
+
+        # energy = [batch size, src len, dec hid dim]
+
+        attention = self.v(energy).squeeze(2)
+
+        # attention= [batch size, src len]
+
+        return F.softmax(attention, dim=1)
