+++ machine_translation/encoders/encoder_gru_attention.py
@@ -1,0 +1,52 @@
+import torch
+import torch.nn as nn
+
+
+class Encoder(nn.Module):
+    def __init__(
+            self,
+            input_dim,
+            emb_dim,
+            enc_hid_dim,
+            dec_hid_dim,
+            dropout
+    ):
+        super().__init__()
+
+        self.embedding = nn.Embedding(input_dim, emb_dim)
+
+        self.rnn = nn.GRU(emb_dim, enc_hid_dim, bidirectional=True)
+
+        self.fc = nn.Linear(enc_hid_dim * 2, dec_hid_dim)
+
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(
+            self,
+            src
+    ):
+        # src = [src len, batch size]
+
+        embedded = self.dropout(self.embedding(src))
+
+        # embedded = [src len, batch size, emb dim]
+
+        outputs, hidden = self.rnn(embedded)
+
+        # outputs = [src len, batch size, hid dim * num directions]
+        # hidden = [n layers * num directions, batch size, hid dim]
+
+        # hidden is stacked [forward_1, backward_1, forward_2, backward_2, ...]
+        # outputs are always from the last layer
+
+        # hidden [-2, :, : ] is the last of the forwards RNN
+        # hidden [-1, :, : ] is the last of the backwards RNN
+
+        # initial decoder hidden is final hidden state of the forwards and backwards
+        #  encoder RNNs fed through a linear layer
+        hidden = torch.tanh(self.fc(torch.cat((hidden[-2, :, :], hidden[-1, :, :]), dim=1)))
+
+        # outputs = [src len, batch size, enc hid dim * 2]
+        # hidden = [batch size, dec hid dim]
+
+        return outputs, hidden
