+++ machine_translation/models/model_seq2seq_gru_attention.py
@@ -4,47 +4,6 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-
-class Encoder(nn.Module):
-    def __init__(self, input_dim, emb_dim, enc_hid_dim, dec_hid_dim, dropout):
-        super().__init__()
-
-        self.embedding = nn.Embedding(input_dim, emb_dim)
-
-        self.rnn = nn.GRU(emb_dim, enc_hid_dim, bidirectional=True)
-
-        self.fc = nn.Linear(enc_hid_dim * 2, dec_hid_dim)
-
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, src):
-        # src = [src len, batch size]
-
-        embedded = self.dropout(self.embedding(src))
-
-        # embedded = [src len, batch size, emb dim]
-
-        outputs, hidden = self.rnn(embedded)
-
-        # outputs = [src len, batch size, hid dim * num directions]
-        # hidden = [n layers * num directions, batch size, hid dim]
-
-        # hidden is stacked [forward_1, backward_1, forward_2, backward_2, ...]
-        # outputs are always from the last layer
-
-        # hidden [-2, :, : ] is the last of the forwards RNN
-        # hidden [-1, :, : ] is the last of the backwards RNN
-
-        # initial decoder hidden is final hidden state of the forwards and backwards
-        #  encoder RNNs fed through a linear layer
-        hidden = torch.tanh(self.fc(torch.cat((hidden[-2, :, :], hidden[-1, :, :]), dim=1)))
-
-        # outputs = [src len, batch size, enc hid dim * 2]
-        # hidden = [batch size, dec hid dim]
-
-        return outputs, hidden
-
-
 class Attention(nn.Module):
     def __init__(self, enc_hid_dim, dec_hid_dim):
         super().__init__()
@@ -76,80 +35,6 @@ class Attention(nn.Module):
         # attention= [batch size, src len]
 
         return F.softmax(attention, dim=1)
-
-
-class Decoder(nn.Module):
-    def __init__(self, output_dim, emb_dim, enc_hid_dim, dec_hid_dim, dropout, attention):
-        super().__init__()
-
-        self.output_dim = output_dim
-        self.attention = attention
-
-        self.embedding = nn.Embedding(output_dim, emb_dim)
-
-        self.rnn = nn.GRU((enc_hid_dim * 2) + emb_dim, dec_hid_dim)
-
-        self.fc_out = nn.Linear((enc_hid_dim * 2) + dec_hid_dim + emb_dim, output_dim)
-
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, input, hidden, encoder_outputs):
-        # input = [batch size]
-        # hidden = [batch size, dec hid dim]
-        # encoder_outputs = [src len, batch size, enc hid dim * 2]
-
-        input = input.unsqueeze(0)
-
-        # input = [1, batch size]
-
-        embedded = self.dropout(self.embedding(input))
-
-        # embedded = [1, batch size, emb dim]
-
-        a = self.attention(hidden, encoder_outputs)
-
-        # a = [batch size, src len]
-
-        a = a.unsqueeze(1)
-
-        # a = [batch size, 1, src len]
-
-        encoder_outputs = encoder_outputs.permute(1, 0, 2)
-
-        # encoder_outputs = [batch size, src len, enc hid dim * 2]
-
-        weighted = torch.bmm(a, encoder_outputs)
-
-        # weighted = [batch size, 1, enc hid dim * 2]
-
-        weighted = weighted.permute(1, 0, 2)
-
-        # weighted = [1, batch size, enc hid dim * 2]
-
-        rnn_input = torch.cat((embedded, weighted), dim=2)
-
-        # rnn_input = [1, batch size, (enc hid dim * 2) + emb dim]
-
-        output, hidden = self.rnn(rnn_input, hidden.unsqueeze(0))
-
-        # output = [seq len, batch size, dec hid dim * n directions]
-        # hidden = [n layers * n directions, batch size, dec hid dim]
-
-        # seq len, n layers and n directions will always be 1 in this decoder, therefore:
-        # output = [1, batch size, dec hid dim]
-        # hidden = [1, batch size, dec hid dim]
-        # this also means that output == hidden
-        assert (output == hidden).all()
-
-        embedded = embedded.squeeze(0)
-        output = output.squeeze(0)
-        weighted = weighted.squeeze(0)
-
-        prediction = self.fc_out(torch.cat((output, weighted, embedded), dim=1))
-
-        # prediction = [batch size, output dim]
-
-        return prediction, hidden.squeeze(0)
 
 
 class Seq2SeqAtt(nn.Module):
