+++ machine_translation/seq2seq/__init__.py
@@ -1,1 +1,3 @@
 from .seq2seq_lstm import Seq2SeqLSTM
+from .seq2seq_gru import Seq2SeqGRU
+from .seq2seq_gru_attention import Seq2SeqGRUAtt
